Feature/134 선점 방식 변경

+++ application/app-item/src/main/java/store/nightmarket/application/appitem/in/dto/ModifyShoppingBasketProductDto.java
@@ -1,6 +1,6 @@
 package store.nightmarket.application.appitem.in.dto;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import lombok.Builder;
 
@@ -8,7 +8,7 @@ public class ModifyShoppingBasketProductDto {
 
 	@Builder
 	public record Request(
-		BigDecimal quantity
+		BigInteger quantity
 	) {
 
 	}
