+++ application/app-item/src/main/java/store/nightmarket/application/appitem/in/dto/ReadShoppingBasketDto.java
@@ -1,6 +1,6 @@
 package store.nightmarket.application.appitem.in.dto;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.UUID;
 
@@ -21,8 +21,8 @@ public class ReadShoppingBasketDto {
 		UUID productVariantId,
 		UUID productPostId,
 		String name,
-		BigDecimal quantity,
-		BigDecimal price,
+		BigInteger quantity,
+		BigInteger price,
 		String imageUrl
 	) {
