+++ application/app-item/src/main/java/store/nightmarket/application/appitem/in/dto/SaveShoppingBasketProductDto.java
@@ -1,6 +1,6 @@
 package store.nightmarket.application.appitem.in.dto;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.UUID;
 
 import lombok.Builder;
@@ -11,8 +11,8 @@ public class SaveShoppingBasketProductDto {
 	public record Request(
 		UUID productVariantId,
 		String name,
-		BigDecimal price,
-		BigDecimal quantity
+		BigInteger price,
+		BigInteger quantity
 	) {
 
 	}
